requirements.txt
pycryptodome==3.17
streamlit==1.16.0
